requirements-dev.txt
black==22.10.0
prospector==1.10.2
google-cloud-dns==0.34.1
google-cloud-resource-manager==1.8.1
dnspython==2.1.0
requests==2.31.0